refactor: rename components to component
