+++ online-store/src/core/components/main/index.ts
@@ -1,4 +1,4 @@
-import Component from '../../templates/components';
+import Component from '../../templates/component';
 
 class Main extends Component {
   renderMain() {
