+++ online-store/src/core/templates/component.ts
@@ -1,4 +1,4 @@
-abstract class Component {
+class Component {
   protected container: HTMLElement;
 
   constructor(tagName: string, className: string) {
